Consistent definition of priority levels
In the specs and the documentation of the spawn_opt family of functions
there is the following definition of priority levels:

	Level = low | normal | high

However, the process_flag/1 function in that module reads:

  process_flag(Flag :: priority, Level) -> OldLevel

    Types:

	Level = OldLevel = priority_level()
	priority_level() = low | normal | high | max

This is clearly inconsistent. Change specs to use the process_level()
type declaration instead of re-defining it in various places.